@@ -74,12 +74,12 @@
     "npm-run-all": "4.1.5",
     "prettier": "1.16.3",
     "prop-types": "15.6.2",
-    "react": "16.7.0",
-    "react-dom": "16.7.0",
+    "react": "16.8.0-alpha.1",
+    "react-dom": "16.8.0-alpha.1",
     "react-relay": "2.0.0",
     "react-router": "4.3.1",
     "react-router-dom": "4.3.1",
-    "react-test-renderer": "16.7.0",
+    "react-test-renderer": "16.8.0-alpha.1",
     "relay-compiler": "2.0.0",
     "storybook-addon-jsx": "6.0.0",
     "xunit-file": "1.0.0"
